requirements.txt
fastapi
uvicorn
pydantic
supabase
python-dotenv
requests
google-genai
groq
qdrant-client
rank_bm25
tqdm
pymupdf
